Added in Lights, cover platform for lights (upper part), and the flip platform trap all in Disco zone.  Lights might need a secondary object for sprites but that can be added if needed.
Also the light "flash" cover is sitting over the disco zone and it linked to all the lights (If needed to be changed use a cube, not a plane).

Note: to remove the ray lines from the lights go into the script and comment out the debug line when ready to run as full game (keep for now for testing and moving).

Added in universal power buttons to ends of Cryo.  Top one for the elevators bottom one for all buttons to other regions.